Changes:
- proper copy of registry
- schema support in foreign keys
- register_default flag in tenant models meta
- deprecate registry attribute of fields

+++ tests/tenancy/test_mt_bulk_create_different_db.py
@@ -15,10 +15,8 @@ from tests.settings import DATABASE_ALTERNATIVE_URL, DATABASE_URL
 pytestmark = pytest.mark.anyio
 
 database = DatabaseTestClient(DATABASE_URL)
-another_db = DatabaseTestClient(DATABASE_ALTERNATIVE_URL)
+another_db = DatabaseTestClient(DATABASE_ALTERNATIVE_URL, use_existing=False)
 models = edgy.Registry(database=database, extra={"another": another_db})
-registry = copy.copy(models)
-registry.database = another_db
 
 
 def time():
@@ -51,30 +49,24 @@ class Product(edgy.Model):
         registry = models
 
 
-@pytest.fixture(autouse=True, scope="module")
+registry = copy.copy(models)
+registry.database = another_db
+
+
+@pytest.fixture(autouse=True, scope="function")
 async def create_test_database():
-    await models.create_all()
-    await registry.create_all()
-    yield
-    await models.drop_all()
-    await registry.drop_all()
+    async with database, another_db:
+        await models.create_all()
+        registry.metadata = models.metadata
+        await registry.create_all(False)
+        yield
+        if not database.drop:
+            await models.drop_all()
+        if not another_db.drop:
+            await registry.drop_all()
 
 
-@pytest.fixture(autouse=True)
-async def rollback_transactions():
-    with database.force_rollback():
-        async with database:
-            yield
-
-
-@pytest.fixture(autouse=True)
-async def rollback_another_db_transactions():
-    with another_db.force_rollback():
-        async with another_db:
-            yield
-
-
-async def test_bulk_create_another_tenant():
+async def test_bulk_create_another_db():
     await Product.query.using_with_db("another").bulk_create(
         [
             {"data": {"foo": 123}, "value": 123.456, "status": StatusEnum.RELEASED},
@@ -89,3 +81,29 @@ async def test_bulk_create_another_tenant():
     others = await Product.query.using_with_db("another").all()
 
     assert len(others) == 2
+
+
+async def test_bulk_create_another_schema_and_db():
+    await registry.schema.create_schema("foo", init_models=True, if_not_exists=True)
+    await registry.create_all(False)
+    try:
+        await Product.query.using_with_db("another", "foo").bulk_create(
+            [
+                {"data": {"foo": 123}, "value": 123.456, "status": StatusEnum.RELEASED},
+                {"data": {"foo": 456}, "value": 456.789, "status": StatusEnum.DRAFT},
+            ]
+        )
+
+        products = await Product.query.all()
+
+        assert len(products) == 0
+
+        products = await Product.query.using_with_db("another").all()
+
+        assert len(products) == 0
+
+        others = await Product.query.using_with_db("another", "foo").all()
+
+        assert len(others) == 2
+    finally:
+        await registry.schema.drop_schema("foo", cascade=True)
